+++ CHANGELOG.md
@@ -55,6 +55,8 @@ handling for NIBI via WNIBI.
 - [#2353](https://github.com/NibiruChain/nibiru/pull/2353) - refactor(oracle): remove dead code from asset registry 
 - [#2371](https://github.com/NibiruChain/nibiru/pull/2371) - feat(evm): fix UnmarshalJSON to accept ASCII hex strings
 - [#2372](https://github.com/NibiruChain/nibiru/pull/2372) - feat(tokenfactory-cli): add CLI commands for set denom functions
+- [#2xxx](https://github.com/NibiruChain/nibiru/pull/2xxx) - feat(evm): support
+WNIBI automatically in the FunToken precompile `sendToBank` and `sendToEvm` methods. 
 
 ### Dependencies
 - Bump `base-x` from 3.0.10 to 3.0.11 ([#2355](https://github.com/NibiruChain/nibiru/pull/2355))